+++ calm/4-apps/src/calm/solid/d/AppModule.scala
@@ -1,0 +1,14 @@
+package calm.solid
+
+trait AppModule extends
+  AuthEntitiesModule with
+  WebEntityModel with
+  AuthCoreModule with
+  CoreModule with
+  WebCoreModule with
+  AuthModule with
+  WebModule
+
+object MainApp extends App with AppModule {
+
+}
